routes: Reuse existing healthcheck instead of a new route
